fix(core): the values document is the values, one per line (KEN-558) [no-changelog]
The stored values a hook's entry carries — every env value and every
header value of its matched registrations — are one document of
stored values, one value per line, and nothing else: no keys, no
braces, no matcher, cwd, url or event. It is absent when the entry
stores none. The hook reading moves from quality/observe.rs into
quality/observe/hook.rs, under the file cap.
