+++ crates/core/src/engine/gate/input.rs
@@ -115,14 +115,14 @@ fn registration(
                 event: event.clone(),
                 matcher: matcher.clone(),
                 command: command.clone(),
-                values: Vec::new(),
+                values: None,
                 script: None,
             },
             None => Content::Hook {
                 event: String::new(),
                 matcher: None,
                 command: location.clone(),
-                values: Vec::new(),
+                values: None,
                 script: None,
             },
         },
@@ -130,7 +130,7 @@ fn registration(
             event: String::new(),
             matcher: None,
             command: location.clone(),
-            values: Vec::new(),
+            values: None,
             script: script.map(|(_, bytes)| String::from_utf8_lossy(bytes).into_owned()),
         },
     };
