+++ crates/core/src/source/browse/safety/input.rs
@@ -51,7 +51,7 @@ pub(super) fn input_for(
             event: String::new(),
             matcher: None,
             command: location.clone(),
-            values: Vec::new(),
+            values: None,
             script: Some(browsed.sealed.read_to_string(path)?),
         },
         _ => Content::Document {
